requirements.txt
google-cloud-aiplatform>=1.36.0
google-cloud-storage>=2.10.0
google-cloud-monitoring>=2.15.0
docker>=6.1.0
python-dotenv>=1.0.0
pyyaml>=6.0.1